move tagOpposites to apf/apfMesh.h
- zoltan/apfZoltan.h: remove tagOpposites declaration.
- zoltan/apfInterElement.cc: remove tagOpposites source.
- apf/apfMesh.h: add tagOpposites declaration.
- reformat documentation comment.
- apf/apfOpposites: add new file with tagOpposites source and helpers.
- apf/CMakeLists.txt: add apfOpposites.cc to SOURCES.
